@@ -10,14 +10,19 @@ import (
 )
 
 const (
-	defaultTitle       = "Health Dashboard | Gatus"
-	defaultDescription = "Gatus is an advanced automated status page that lets you monitor your applications and configure alerts to notify you if there's an issue"
-	defaultHeader      = "Gatus"
-	defaultLogo        = ""
-	defaultLink        = ""
-	defaultCustomCSS   = ""
-	defaultSortBy      = "name"
-	defaultFilterBy    = "none"
+	defaultTitle                = "Health Dashboard | Gatus"
+	defaultDescription          = "Gatus is an advanced automated status page that lets you monitor your applications and configure alerts to notify you if there's an issue"
+	defaultHeader               = "Gatus"
+	defaultDashboardHeading     = "Health Dashboard"
+	defaultDashboardSubheading  = "Monitor the health of your endpoints in real-time"
+	defaultLogo                 = ""
+	defaultLink                 = ""
+	defaultFavicon              = "/favicon.ico"
+	defaultFavicon16            = "/favicon-16x16.png"
+	defaultFavicon32            = "/favicon-32x32.png"
+	defaultCustomCSS            = ""
+	defaultSortBy               = "name"
+	defaultFilterBy             = "none"
 )
 
 var (
@@ -30,17 +35,19 @@ var (
 
 // Config is the configuration for the UI of Gatus
 type Config struct {
-	Title           string   `yaml:"title,omitempty"`             // Title of the page
-	Description     string   `yaml:"description,omitempty"`       // Meta description of the page
-	Header          string   `yaml:"header,omitempty"`            // Header is the text at the top of the page
-	Logo            string   `yaml:"logo,omitempty"`              // Logo to display on the page
-	Link            string   `yaml:"link,omitempty"`              // Link to open when clicking on the logo
-	Buttons         []Button `yaml:"buttons,omitempty"`           // Buttons to display below the header
-	CustomCSS       string   `yaml:"custom-css,omitempty"`        // Custom CSS to include in the page
-	DarkMode        *bool    `yaml:"dark-mode,omitempty"`         // DarkMode is a flag to enable dark mode by default
-	DefaultSortBy   string   `yaml:"default-sort-by,omitempty"`   // DefaultSortBy is the default sort option ('name', 'group', 'health')
-	DefaultFilterBy string   `yaml:"default-filter-by,omitempty"` // DefaultFilterBy is the default filter option ('none', 'failing', 'unstable')
-
+	Title                   string   `yaml:"title,omitempty"`                  // Title of the page
+	Description             string   `yaml:"description,omitempty"`            // Meta description of the page
+	DashboardHeading        string   `yaml:"dashboard-heading,omitempty"`      // Dashboard Title between header and endpoints
+	DashboardSubheading     string   `yaml:"dashboard-subheading,omitempty"`   // Dashboard Description between header and endpoints
+	Header                  string   `yaml:"header,omitempty"`                 // Header is the text at the top of the page
+	Logo                    string   `yaml:"logo,omitempty"`                   // Logo to display on the page
+	Link                    string   `yaml:"link,omitempty"`                   // Link to open when clicking on the logo
+	Favicon                 Favicon  `yaml:"favicon,omitempty"`                // Favourite icon to display in web browser tab or address bar
+	Buttons                 []Button `yaml:"buttons,omitempty"`                // Buttons to display below the header
+	CustomCSS               string   `yaml:"custom-css,omitempty"`             // Custom CSS to include in the page
+	DarkMode                *bool    `yaml:"dark-mode,omitempty"`              // DarkMode is a flag to enable dark mode by default
+	DefaultSortBy           string   `yaml:"default-sort-by,omitempty"`        // DefaultSortBy is the default sort option ('name', 'group', 'health')
+	DefaultFilterBy         string   `yaml:"default-filter-by,omitempty"`      // DefaultFilterBy is the default filter option ('none', 'failing', 'unstable')
 	//////////////////////////////////////////////
 	// Non-configurable - used for UI rendering //
 	//////////////////////////////////////////////
@@ -68,11 +75,19 @@ func (btn *Button) Validate() error {
 	return nil
 }
 
+type Favicon struct {
+	Default   string `yaml:"default,omitempty"`   // URL or path to default favourite icon.
+	Size16x16 string `yaml:"size16x16,omitempty"` // URL or path to favourite icon for 16x16 size.
+	Size32x32 string `yaml:"size32x32,omitempty"` // URL or path to favourite icon for 32x32 size.
+}
+
 // GetDefaultConfig returns a Config struct with the default values
 func GetDefaultConfig() *Config {
 	return &Config{
 		Title:                  defaultTitle,
 		Description:            defaultDescription,
+		DashboardHeading:       defaultDashboardHeading,
+		DashboardSubheading:    defaultDashboardSubheading,
 		Header:                 defaultHeader,
 		Logo:                   defaultLogo,
 		Link:                   defaultLink,
@@ -81,6 +96,11 @@ func GetDefaultConfig() *Config {
 		DefaultSortBy:          defaultSortBy,
 		DefaultFilterBy:        defaultFilterBy,
 		MaximumNumberOfResults: storage.DefaultMaximumNumberOfResults,
+		Favicon: Favicon{
+			Default:   defaultFavicon,
+			Size16x16: defaultFavicon16,
+			Size32x32: defaultFavicon32,
+		},
 	}
 }
 
@@ -91,6 +111,12 @@ func (cfg *Config) ValidateAndSetDefaults() error {
 	}
 	if len(cfg.Description) == 0 {
 		cfg.Description = defaultDescription
+	}
+	if len(cfg.DashboardHeading) == 0 {
+		cfg.DashboardHeading = defaultDashboardHeading
+	}
+	if len(cfg.DashboardSubheading) == 0 {
+		cfg.DashboardSubheading = defaultDashboardSubheading
 	}
 	if len(cfg.Header) == 0 {
 		cfg.Header = defaultHeader
@@ -116,6 +142,15 @@ func (cfg *Config) ValidateAndSetDefaults() error {
 		cfg.DefaultFilterBy = defaultFilterBy
 	} else if cfg.DefaultFilterBy != "none" && cfg.DefaultFilterBy != "failing" && cfg.DefaultFilterBy != "unstable" {
 		return ErrInvalidDefaultFilterBy
+	}
+	if len(cfg.Favicon.Default) == 0 {
+		cfg.Favicon.Default = defaultFavicon
+	}
+	if len(cfg.Favicon.Size16x16) == 0 {
+		cfg.Favicon.Size16x16 = defaultFavicon16
+	}
+	if len(cfg.Favicon.Size32x32) == 0 {
+		cfg.Favicon.Size32x32 = defaultFavicon32
 	}
 	for _, btn := range cfg.Buttons {
 		if err := btn.Validate(); err != nil {
