@@ -46,7 +46,7 @@
               {{ endpoint.success ? '✓' : '✗' }}
             </span>
             <span class="truncate">{{ endpoint.name }}</span>
-            <span class="text-muted-foreground">({{ (endpoint.duration / 1000000).toFixed(0) }}ms)</span>
+            <span class="text-muted-foreground">({{ Math.trunc(endpoint.duration / 1000000) }}ms)</span>
           </div>
           <div v-if="result.endpointResults.length > 5" class="text-xs text-muted-foreground">
             ... and {{ result.endpointResults.length - 5 }} more
@@ -60,7 +60,7 @@
           {{ isSuiteResult ? 'Total Duration' : 'Response Time' }}
         </div>
         <div class="font-mono text-xs">
-          {{ isSuiteResult ? (result.duration / 1000000).toFixed(0) : (result.duration / 1000000).toFixed(0) }}ms
+          {{ Math.trunc(result.duration / 1000000) }}ms
         </div>
       </div>
       
@@ -95,7 +95,6 @@
 </template>
 
 <script setup>
-/* eslint-disable no-undef */
 import { ref, watch, nextTick, computed, onMounted, onUnmounted } from 'vue'
 import { useRoute } from 'vue-router'
 import { prettifyTimestamp } from '@/utils/time'
