@@ -35,7 +35,7 @@
                 <CardTitle class="text-sm font-medium text-muted-foreground">Avg Response Time</CardTitle>
               </CardHeader>
               <CardContent>
-                <div class="text-2xl font-bold">{{ pageAverageResponseTime }}ms</div>
+                <div class="text-2xl font-bold">{{ pageAverageResponseTime }}</div>
               </CardContent>
             </Card>
 
@@ -66,7 +66,7 @@
                   <Button 
                     variant="ghost" 
                     size="icon"
-                    @click="showAverageResponseTime = !showAverageResponseTime"
+                    @click="toggleShowAverageResponseTime"
                     :title="showAverageResponseTime ? 'Show min-max response time' : 'Show average response time'"
                   >
                     <Activity v-if="showAverageResponseTime" class="h-5 w-5" />
@@ -89,13 +89,13 @@
                 <EndpointCard 
                   v-if="endpointStatus"
                   :endpoint="endpointStatus"
-                  :maxResults="50"
+                  :maxResults="resultPageSize"
                   :showAverageResponseTime="showAverageResponseTime"
                   @showTooltip="showTooltip"
                   class="border-0 shadow-none bg-transparent p-0"
                 />
                 <div v-if="endpointStatus && endpointStatus.key" class="pt-4 border-t">
-                  <Pagination @page="changePage" :numberOfResultsPerPage="50" :currentPageProp="currentPage" />
+                  <Pagination @page="changePage" :numberOfResultsPerPage="resultPageSize" :currentPageProp="currentPage" />
                 </div>
               </div>
             </CardContent>
@@ -201,7 +201,6 @@
 </template>
 
 <script setup>
-/* eslint-disable no-undef */
 import { ref, computed, onMounted } from 'vue'
 import { useRouter, useRoute } from 'vue-router'
 import { ArrowLeft, RefreshCw, ArrowUpCircle, ArrowDownCircle, PlayCircle, Activity, Timer } from 'lucide-vue-next'
@@ -213,7 +212,6 @@ import Settings from '@/components/Settings.vue'
 import Pagination from '@/components/Pagination.vue'
 import Loading from '@/components/Loading.vue'
 import ResponseTimeChart from '@/components/ResponseTimeChart.vue'
-import { SERVER_URL } from '@/main.js'
 import { generatePrettyTimeAgo, generatePrettyTimeDifference } from '@/utils/time'
 
 const router = useRouter()
@@ -224,10 +222,10 @@ const endpointStatus = ref(null) // For paginated historical data
 const currentStatus = ref(null) // For current/latest status (always page 1)
 const events = ref([])
 const currentPage = ref(1)
+const resultPageSize = 50
 const showResponseTimeChartAndBadges = ref(false)
-const showAverageResponseTime = ref(false)
+const showAverageResponseTime = ref(localStorage.getItem('gatus:show-average-response-time') !== 'false')
 const selectedChartDuration = ref('24h')
-const serverUrl = SERVER_URL === '.' ? '..' : SERVER_URL
 const isRefreshing = ref(false)
 
 const latestResult = computed(() => {
@@ -247,6 +245,11 @@ const hostname = computed(() => {
   return latestResult.value?.hostname || null
 })
 
+const toggleShowAverageResponseTime = () => {
+  showAverageResponseTime.value = !showAverageResponseTime.value
+  localStorage.setItem('gatus:show-average-response-time', showAverageResponseTime.value ? 'true' : 'false')
+}
+
 const pageAverageResponseTime = computed(() => {
   // Use endpointStatus for current page's average response time
   if (!endpointStatus.value || !endpointStatus.value.results || endpointStatus.value.results.length === 0) {
@@ -261,7 +264,7 @@ const pageAverageResponseTime = computed(() => {
     }
   }
   if (count === 0) return 'N/A'
-  return Math.round(total / count / 1000000)
+  return `${Math.round(total / count / 1000000)}ms`
 })
 
 const pageResponseTimeRange = computed(() => {
@@ -274,17 +277,17 @@ const pageResponseTimeRange = computed(() => {
   let hasData = false
   
   for (const result of endpointStatus.value.results) {
-    if (result.duration) {
-      const durationMs = result.duration / 1000000
-      min = Math.min(min, durationMs)
-      max = Math.max(max, durationMs)
+    const duration = result.duration
+    if (duration) {
+      min = Math.min(min, duration)
+      max = Math.max(max, duration)
       hasData = true
     }
   }
   
   if (!hasData) return 'N/A'
-  const minMs = Math.round(min)
-  const maxMs = Math.round(max)
+  const minMs = Math.trunc(min / 1000000)
+  const maxMs = Math.trunc(max / 1000000)
   // If min and max are the same, show single value
   if (minMs === maxMs) {
     return `${minMs}ms`
@@ -304,7 +307,7 @@ const lastCheckTime = computed(() => {
 const fetchData = async () => {
   isRefreshing.value = true
   try {
-    const response = await fetch(`${serverUrl}/api/v1/endpoints/${route.params.key}/statuses?page=${currentPage.value}&pageSize=50`, {
+    const response = await fetch(`/api/v1/endpoints/${route.params.key}/statuses?page=${currentPage.value}&pageSize=${resultPageSize}`, {
       credentials: 'include'
     })
     
@@ -385,15 +388,15 @@ const prettifyTimestamp = (timestamp) => {
 }
 
 const generateHealthBadgeImageURL = () => {
-  return `${serverUrl}/api/v1/endpoints/${endpointStatus.value.key}/health/badge.svg`
+  return `/api/v1/endpoints/${endpointStatus.value.key}/health/badge.svg`
 }
 
 const generateUptimeBadgeImageURL = (duration) => {
-  return `${serverUrl}/api/v1/endpoints/${endpointStatus.value.key}/uptimes/${duration}/badge.svg`
+  return `/api/v1/endpoints/${endpointStatus.value.key}/uptimes/${duration}/badge.svg`
 }
 
 const generateResponseTimeBadgeImageURL = (duration) => {
-  return `${serverUrl}/api/v1/endpoints/${endpointStatus.value.key}/response-times/${duration}/badge.svg`
+  return `/api/v1/endpoints/${endpointStatus.value.key}/response-times/${duration}/badge.svg`
 }
 
 onMounted(() => {
